Create open_workspace.ps1
Loading script into a repo. Need to develop an installer and a way to inject path locations. Need to develop a URL/Window extractor that is able to grab current urls/window information and save to a file so that this can read it later. 

Need to see the proper set up and management for this kind of project.