Update open_workspace.ps1 for comma bug and small array bug
Currently, workspace.txt has urls separated by new lines with commas before each new line, except for the final url. This might change if we can replace the current extension that generates these. 

But when extracting, Get-Content reads a file line by line, strings are generated at new line characters. So the comma needed trimming to avoid some urls breaking.

Further, if you only had one url in a group before a window was launched, the array identity would be dropped due to powershell trying to be smart. But we relied on .Clear() to reset the url array, which is an array method and not a string method. Thus it would fail, but the script would continue. So this would compound, because we would loop through the file, but keep encountering the first delimiter and fail to .Clear() again and again until we finished processing the file.

Instead of .Clear(), we reset the array to an empty array. Bugs fixed.